Add checks/tests for the build process

+++ scripts/build-shared.js
@@ -10,6 +10,15 @@ const BUILD_DIR = path.join(__dirname, '../intermediate-builds/shared/')
 if (process.env.NODE_ENV !== 'production') {
   throw new Error('Builds must have NODE_ENV=production')
 }
+const browserName = process.env.BROWSER
+if (!browserName) {
+  throw new Error('The environment variable process.env.BROWSER must be set.')
+}
+if (['chrome', 'firefox'].indexOf(browserName) < 0) {
+  throw new Error(
+    `The environment variable process.env.BROWSER must be one of: "chrome", "firefox". Received: "${browserName}".`
+  )
+}
 
 // Empty the build directory.
 fs.emptyDirSync(BUILD_DIR)
